+++ react/TickIcon/TickIcon.js
@@ -1,3 +1,3 @@
-import { iconGenerator } from '../Icons/Icons';
+import iconGenerator from '../private/IconsGenerator/IconsGenerator';
 
 export default iconGenerator('TickIcon');
