+++ react/private/IconsGenerator/IconsGenerator.less
@@ -1,5 +1,3 @@
-// TickCircleIcon
-
 .TickCircleIcon.filled,
 .HeartIcon.filled {
   animation-duration: 0.4s;
