+++ react/private/IconsGenerator/svg/ThumbsUpIcon.js
